Output: Change audio to use synthetic timecode.
Video output from OBS is essentially clocked by OBS, and the audio is
essentially clocked to that, so instead of using the timestamps of the
audio frames which seem unreliable, use the timecode synthesize
functionality from the NDI library. This should mean that the audio
stays synced with the video.